model/requirements.txt
torch
torchvision
segmentation-models-pytorch
albumentations
opencv-python
matplotlib
pillow
numpy
pandas
scikit-learn
